hubAPI - check page, limit, offset, page_size, sort, ordering, order_by
throws if the wrong param is used based on the url

Issue: AAH-3007
