+++ frontend/hub/common/api/query-string.ts
@@ -1,0 +1,33 @@
+const sortKeys = {
+  '/pulp/api/v3/': 'ordering',
+  '/pulp/api/v3/pulp_container/namespaces/': 'sort',
+  '/v1/imports/': 'order_by',
+  '/v1/roles/': 'order_by',
+  '/_ui/v1/': 'sort',
+  '/v3/plugin/ansible/search/collection-versions/': 'order_by',
+};
+
+const pageKeys = {
+  '/v1/imports/': 'page',
+  '/v1/namespaces/': 'page',
+  '/v1/roles/': 'page',
+  '/_ui/v1/': 'offset',
+};
+
+export function url2keys(url: string): { sortKey: string; pageKey: string } {
+  let sortKey = 'sort';
+  Object.entries(sortKeys).forEach(([k, v]) => {
+    if (url.includes(k)) {
+      sortKey = v;
+    }
+  });
+
+  let pageKey = 'offset';
+  Object.entries(pageKeys).forEach(([k, v]) => {
+    if (url.includes(k)) {
+      pageKey = v;
+    }
+  });
+
+  return { pageKey, sortKey };
+}
