+++ frontend/hub/common/useHubView.tsx
@@ -11,6 +11,7 @@ import {
 import { useFetcher } from '../../common/crud/Data';
 import { RequestError } from '../../common/crud/RequestError';
 import { QueryParams, getQueryString, serverlessURL } from './api/hub-api-utils';
+import { url2keys } from './api/query-string';
 
 export interface HubItemsResponse<T extends object> {
   meta: {
@@ -60,40 +61,6 @@ function deconstruct<T extends object>(
       pageItems: data?.results,
     };
   }
-}
-
-const sortKeys = {
-  '/pulp/api/v3/': 'ordering',
-  '/pulp/api/v3/pulp_container/namespaces/': 'sort',
-  '/v1/imports/': 'order_by',
-  '/v1/roles/': 'order_by',
-  '/_ui/v1/': 'sort',
-  '/v3/plugin/ansible/search/collection-versions/': 'order_by',
-};
-
-const pageKeys = {
-  '/v1/imports/': 'page',
-  '/v1/namespaces/': 'page',
-  '/v1/roles/': 'page',
-  '/_ui/v1/': 'offset',
-};
-
-function url2keys(url: string): { sortKey: string; pageKey: string } {
-  let sortKey = 'sort';
-  Object.entries(sortKeys).forEach(([k, v]) => {
-    if (url.includes(k)) {
-      sortKey = v;
-    }
-  });
-
-  let pageKey = 'offset';
-  Object.entries(pageKeys).forEach(([k, v]) => {
-    if (url.includes(k)) {
-      pageKey = v;
-    }
-  });
-
-  return { pageKey, sortKey };
 }
 
 export function useHubView<T extends object>({
